requirements.txt
cryptography==38.0.4
python-dotenv==1.0.0
pyfiglet==0.8.post1